requirements.txt
selenium
pandas
requests
ortools
scikit-learn
python-dotenv